+++ pkg/webhook/webhook-server.go
@@ -1,4 +1,4 @@
-// Copyright © 2019-2020 Talend - www.talend.com
+// Copyright © 2019-2021 Talend - www.talend.com
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -20,11 +20,10 @@ import (
 	"io/ioutil"
 	"net/http"
 	cfg "talend/vault-sidecar-injector/pkg/config"
-	ctx "talend/vault-sidecar-injector/pkg/context"
 	m "talend/vault-sidecar-injector/pkg/mode"
 
-	"k8s.io/api/admission/v1beta1"
-	corev1 "k8s.io/api/core/v1"
+	admv1 "k8s.io/api/admission/v1"
+	admv1beta1 "k8s.io/api/admission/v1beta1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/runtime/serializer"
@@ -65,142 +64,105 @@ func New(config *cfg.VSIConfig, server *http.Server) *VaultInjector {
 	}
 }
 
-// Create mutation patch for resources
-func (vaultInjector *VaultInjector) createPatch(pod *corev1.Pod, annotations map[string]string) ([]byte, error) {
-
-	patchPodSpec, err := vaultInjector.updatePodSpec(pod)
-	if err != nil {
-		return nil, err
-	}
-
-	var patch []ctx.PatchOperation
-
-	patch = append(patch, patchPodSpec...)
-	patch = append(patch, updateAnnotation(pod.Annotations, annotations)...)
-
-	return json.Marshal(patch)
-}
-
-// Main mutation process
-func (vaultInjector *VaultInjector) mutate(ar *v1beta1.AdmissionReview) *v1beta1.AdmissionResponse {
-	req := ar.Request
-	var pod corev1.Pod
-	var podName string
-	var podNamespace string
-
-	if klog.V(5) { // enabled by providing '-v=5' at least
-		klog.Infof("Request=%+v", req)
-	}
-
-	if err := json.Unmarshal(req.Object.Raw, &pod); err != nil {
-		klog.Errorf("Could not unmarshal raw object: %v", err)
-		return &v1beta1.AdmissionResponse{
-			Result: &metav1.Status{
-				Message: err.Error(),
-			},
-		}
-	}
-
-	if klog.V(5) { // enabled by providing '-v=5' at least
-		klog.Infof("Pod=%+v", pod)
-	}
-
-	if pod.Name == "" {
-		podName = pod.GenerateName
-	} else {
-		podName = pod.Name
-	}
-
-	if pod.Namespace == "" {
-		podNamespace = metav1.NamespaceDefault
-	} else {
-		podNamespace = pod.Namespace
-	}
-
-	klog.Infof("AdmissionReview for GroupVersionKind=%+v, Namespace=%v Name=%v (%v) UID=%v patchOperation=%v UserInfo=%+v",
-		req.Kind, req.Namespace, req.Name, podName, req.UID, req.Operation, req.UserInfo)
-
-	// Determine whether to perform mutation
-	if !mutationRequired(ignoredNamespaces, vaultInjector.VaultInjectorAnnotationsFQ, &pod.ObjectMeta) {
-		klog.Infof("Skipping mutation for %s/%s due to policy check", podNamespace, podName)
-		return &v1beta1.AdmissionResponse{
-			Allowed: true,
-		}
-	}
-
-	annotations := map[string]string{vaultInjector.VaultInjectorAnnotationsFQ[ctx.VaultInjectorAnnotationStatusKey]: ctx.VaultInjectorStatusInjected}
-	patchBytes, err := vaultInjector.createPatch(&pod, annotations)
-	if err != nil {
-		return &v1beta1.AdmissionResponse{
-			Allowed: false,
-			Result: &metav1.Status{
-				Message: err.Error(),
-			},
-		}
-	}
-
-	klog.Infof("AdmissionResponse: patch=%v\n", string(patchBytes))
-	return &v1beta1.AdmissionResponse{
-		Allowed: true,
-		Patch:   patchBytes,
-		PatchType: func() *v1beta1.PatchType {
-			pt := v1beta1.PatchTypeJSONPatch
-			return &pt
-		}(),
-	}
-}
-
 // Serve method for webhook server
 func (vaultInjector *VaultInjector) Serve(w http.ResponseWriter, r *http.Request) {
 	var body []byte
+
+	if klog.V(5) { // enabled by providing '-v=5' at least
+		klog.Infof("HTTP Request=%+v", r)
+	}
+
 	if r.Body != nil {
 		if data, err := ioutil.ReadAll(r.Body); err == nil {
 			body = data
 		}
 	}
+
 	if len(body) == 0 {
-		klog.Error("empty body")
-		http.Error(w, "empty body", http.StatusBadRequest)
+		klog.Error("Empty body")
+		http.Error(w, "Empty body", http.StatusBadRequest)
 		return
 	}
 
-	// Verify the content type is accurate
 	contentType := r.Header.Get("Content-Type")
 	if contentType != "application/json" {
 		klog.Errorf("Content-Type=%s, expect application/json", contentType)
-		http.Error(w, "invalid Content-Type, expect `application/json`", http.StatusUnsupportedMediaType)
+		http.Error(w, "Invalid Content-Type, expect `application/json`", http.StatusUnsupportedMediaType)
 		return
 	}
 
-	var admissionResponse *v1beta1.AdmissionResponse
-	ar := v1beta1.AdmissionReview{}
-	if _, _, err := deserializer.Decode(body, nil, &ar); err != nil {
+	rawAR, gvk, err := deserializer.Decode(body, nil, nil)
+	if err != nil {
 		klog.Errorf("Can't decode body: %v", err)
-		admissionResponse = &v1beta1.AdmissionResponse{
-			Result: &metav1.Status{
-				Message: err.Error(),
-			},
+		http.Error(w, fmt.Sprintf("Cannot decode body: %v", err), http.StatusBadRequest)
+		return
+	}
+
+	// Webhook can currently receive either v1 or v1beta1 AdmissionReview objects.
+	// v1 is the default, internal version in use by VSI (v1beta1 support will be removed).
+	// If v1beta1 is received, it will be converted into v1 and back to v1beta1 in response
+	// (as spec states response should use same version as request)
+	arInVersion := admv1.SchemeGroupVersion
+	arIn, isV1 := rawAR.(*admv1.AdmissionReview)
+	if !isV1 {
+		arInVersion = admv1beta1.SchemeGroupVersion
+		arInv1beta1, isv1beta1 := rawAR.(*admv1beta1.AdmissionReview)
+		if !isv1beta1 {
+			klog.Errorf("Unsupported AdmissionReview version %v", gvk.Version)
+			http.Error(w, fmt.Sprintf("Unsupported AdmissionReview version %v", gvk.Version), http.StatusBadRequest)
+			return
+		}
+
+		if klog.V(5) { // enabled by providing '-v=5' at least
+			klog.Infof("Received AdmissionReview '%v' Request=%+v", arInv1beta1.GroupVersionKind(), arInv1beta1.Request)
+		}
+
+		// Convert v1beta1 to v1
+		arIn = &admv1.AdmissionReview{}
+		arIn.SetGroupVersionKind(admv1.SchemeGroupVersion.WithKind("AdmissionReview"))
+		Convert_v1beta1_AdmissionReview_To_admission_AdmissionReview(arInv1beta1, arIn)
+
+		if klog.V(5) { // enabled by providing '-v=5' at least
+			klog.Infof("Converted AdmissionReview '%v' Request=%+v", arIn.GroupVersionKind(), arIn.Request)
 		}
 	} else {
-		admissionResponse = vaultInjector.mutate(&ar)
-	}
-
-	admissionReview := v1beta1.AdmissionReview{}
-	if admissionResponse != nil {
-		admissionReview.Response = admissionResponse
-		if ar.Request != nil {
-			admissionReview.Response.UID = ar.Request.UID
+		if klog.V(5) { // enabled by providing '-v=5' at least
+			klog.Infof("Received AdmissionReview '%v' Request=%+v", arIn.GroupVersionKind(), arIn.Request)
 		}
 	}
 
-	resp, err := json.Marshal(admissionReview)
+	arOut := &admv1.AdmissionReview{}
+	arOut.SetGroupVersionKind(admv1.SchemeGroupVersion.WithKind("AdmissionReview"))
+	arOut.Response = vaultInjector.mutate(arIn)
+
+	var returnedAR interface{}
+	returnedAR = arOut
+
+	// If v1 received
+	if arInVersion.Version == admv1.SchemeGroupVersion.Version {
+		if klog.V(5) { // enabled by providing '-v=5' at least
+			klog.Infof("Returned AdmissionReview '%v' Response=%+v", arOut.GroupVersionKind(), arOut.Response)
+		}
+	} else { // If v1beta1 received: convert v1 to v1beta1
+		arOutv1beta1 := &admv1beta1.AdmissionReview{}
+		arOutv1beta1.SetGroupVersionKind(admv1beta1.SchemeGroupVersion.WithKind("AdmissionReview"))
+		Convert_admission_AdmissionReview_To_v1beta1_AdmissionReview(arOut, arOutv1beta1)
+		returnedAR = arOutv1beta1
+
+		if klog.V(5) { // enabled by providing '-v=5' at least
+			klog.Infof("Returned AdmissionReview '%v' Response=%+v", arOutv1beta1.GroupVersionKind(), arOutv1beta1.Response)
+		}
+	}
+
+	response, err := json.Marshal(returnedAR)
 	if err != nil {
 		klog.Errorf("Can't encode response: %v", err)
-		http.Error(w, fmt.Sprintf("could not encode response: %v", err), http.StatusInternalServerError)
+		http.Error(w, fmt.Sprintf("Cannot encode response: %v", err), http.StatusInternalServerError)
 	}
-	klog.Infof("Ready to write reponse ...")
-	if _, err := w.Write(resp); err != nil {
+
+	klog.Infof("Write reponse ...")
+	if _, err := w.Write(response); err != nil {
 		klog.Errorf("Can't write response: %v", err)
-		http.Error(w, fmt.Sprintf("could not write response: %v", err), http.StatusInternalServerError)
+		http.Error(w, fmt.Sprintf("Cannot write response: %v", err), http.StatusInternalServerError)
 	}
 }